refactor: split out the save credentials section into its own file
